iterative changes to GLF files; also a test of checking-in over sshfs.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@850 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/test/org/broadinstitute/sting/utils/glf/LikelihoodObjectTest.java
@@ -80,7 +80,7 @@ public class LikelihoodObjectTest extends BaseTest {
         assertTrue(mLO.likelihood.size() == LikelihoodObject.GENOTYPE.values().length);
 
         int index = 0;
-        int[] ret = mLO.toByteArray();
+        short[] ret = mLO.toByteArray();
         for (index = 0; index < ret.length; index++) {
             assertTrue(ray[index] == ret[index]);
         }
@@ -95,7 +95,7 @@ public class LikelihoodObjectTest extends BaseTest {
         assertTrue(mLO.likelihood.size() == LikelihoodObject.GENOTYPE.values().length);
 
         int index = 0;
-        int[] ret = mLO.toByteArray();
+        short[] ret = mLO.toByteArray();
         for (index = 0; index < ret.length; index++) {
             assertTrue(ret[index] == 128);
         }
